requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
qrcode==7.4.2
Pillow==10.4.0
Werkzeug==3.0.3
reportlab==4.0.7
gunicorn==21.2.0
python-dotenv==1.0.0


